Add files via upload
This is the Winner 2015a but I made a few changes and it seems to be winning more efficiently now. I altered the defensive assignment so there isn't always a defender, which makes the game go by faster as well as seems to use the enemies sprites dying to our advantage.